pyproject.toml
[project]
name = "agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "hasnainbusiness204@gmail.com", email = "163559600+Hasnain-Software-Developer@users.noreply.github.com" }
]
requires-python = ">=3.12"
dependencies = [
    "langchain-google-genai>=2.0.9",
    "langgraph>=0.2.74",
    "python-dotenv>=1.0.1",
]

[project.scripts]
agents = "agents:main"
orch = "agents.orch:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
